// patches/patch-004-inventory-ledger-core/P04_inventory_ledger_core/files/src/shared/types.ts
export type NavPage =
  | "dashboard"
  | "catalog"
  | "inventory"
  | "business"
  | "settings"
  | "data-center";